requirements.txt
aiosqlite
asyncio
fastapi
httpx
pandas
psutil
magika
python-decouple
six==1.16.0
sqlalchemy
starlette
uvicorn[standard]
uvloop
zstandard
